feat(example_extend_cli.py): import LabelParser from pylo.cli.NativeParsers Add an import statement for LabelParser from pylo.cli.NativeParsers. This allows the use of LabelParser in the code.
feat(example_extend_cli.py): add support for sorting workloads by name
feat(example_extend_cli.py): add support for built-in parser to filter workloads by environment label
This commit adds two new features to the CLI command. The first feature is the ability to sort workloads by name. This is achieved by adding a new argument to the argparse parser called '--sort-by-name'. The second feature is the ability to filter workloads by environment label using a built-in parser. The built-in parser is defined as a dataclass called 'MyBuiltInParser' and is used to parse the '--filter-env-label' argument. The results of the parser are then used to filter the workloads that are displayed.

feat(example_extend_cli.py): add support for custom built-in parsers
This commit adds support for custom built-in parsers in the `Command` class. The `native_parsers_as_class` parameter can now be used to specify a custom built-in parser. In this case, the `MyBuiltInParser()` class is used as the custom built-in parser. This allows for greater flexibility in parsing command line arguments.