refactor(LabelStore.py): rename find_label_by_name method to find_object_by_name and add allow_label_group and allow_label parameters The find_label_by_name method has been renamed to find_object_by_name to better reflect its functionality. Two new parameters have been added to allow_label_group and allow_label to allow for more flexibility in the search. If allow_label_group is set to False, the search will ignore label groups. If allow_label is set to False, the search will ignore labels. This allows for more fine-grained control over the search.
feat(LabelStore.py): add find_label_by_name method
The find_label_by_name method allows for finding a label by its name. It takes in a name parameter which can be a string or a list of strings. If case_sensitive is False, the search is case-insensitive. If missing_labels_names is not None, it will be filled with the names of the labels not found. If raise_exception_if_not_found is True, an exception will be raised if a label is not found. If a label is not found, None will be returned in the list. This method is added to the LabelStore class.